Move storage backends to their own folder
