+++ src/backends/dummy.js
@@ -1,3 +1,7 @@
+export function makeNodeFolder (path) {
+  throw new Error('Not available on React Native')
+}
+
 export function makeReactNativeFolder (path) {
   throw new Error('Not available on node.js')
 }
