--- src/nodeDummy.js
@@ -1,3 +1,0 @@
-export function makeNodeFolder (path) {
-  throw new Error('Not available on React Native')
-}
